requirements.txt
Flask
flask_sock
python-dotenv
mysql-connector-python
cryptography
